+++ libnopegl/src/ngpu/vulkan/ctx_vk.c
@@ -862,10 +862,7 @@ static int vk_init(struct ngpu_ctx *s)
                +  10 * VK_API_VERSION_MINOR(vk->api_version);
     s->language_version = 450;
 
-    s->features = NGPU_FEATURE_COMPUTE |
-                  NGPU_FEATURE_IMAGE_LOAD_STORE |
-                  NGPU_FEATURE_STORAGE_BUFFER |
-                  NGPU_FEATURE_BUFFER_MAP_PERSISTENT;
+    s->features = NGPU_FEATURE_BUFFER_MAP_PERSISTENT;
 
     const VkPhysicalDeviceLimits *limits = &vk->phy_device_props.limits;
     s->limits.max_vertex_attributes              = get_max_vertex_attributes(limits);